requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
databases[aiosqlite]>=0.5.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
google-genai>=1.28.0
google>=3.0.0
numpy>=2.3.0